Mapping and IHttpActionResult Methods

+++ ProductProject/Models/Genre.cs
@@ -8,8 +8,7 @@ namespace ProductProject.Models
 {
     public class Genre
     {
-        public int Id { get; set; }
-
+        public byte Id { get; set; }
         [Required]
         [StringLength(255)]
         public string Name { get; set; }
